requirements.txt
pyspark==3.5.6
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
python-dotenv>=0.21.0
minio>=7.1.0
psycopg2-binary>=2.9.5
kafka-python>=2.0.2
